// frontend/src/api/generated/models/jobPacketResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface JobPacketResponse {
  id: string;
  production_order_id: string;
  packet_number: string;
  quantity: number;
  current_stage_id?: string | null;
  current_stage_name?: string | null;
  assigned_operator_id?: string | null;
  status: string;
}